fix (list_sub_domains): fix output.
Output now correctly handles unmanaged subdomains that exist in the database.
Previously, if a subdomain was in the local DB it was treated as if it was managed, even if it wasn't managed.
As a result, both 1) cataloged and unmanaged subdomains and 2) uncataloged subdomains (i.e. A records associated with the domain with no corresponding entries in the DB) are treated the same.